feat: 项目部署配置 + Docker构建 + README完善
- Dockerfile + docker-compose + .dockerignore
- 生产环境配置 .env.production
- init-deploy.bat 一键生成部署目录
- README 更新（启动/部署/接口文档）
- 清理 .pyc / __pycache__ / 测试文件
- 删除无用路由和DAO文件
- 请求数据校验规则完善
- 百炼识别输出JSON格式
- Qdrant连接池优化
- 配置校验（model表/枚举/长度限制）
- models分页开放给登录用户

+++ app/models/block.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String
+from sqlalchemy import DateTime, Float, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models import Base
@@ -12,9 +12,9 @@ class Block(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     question_id: Mapped[int] = mapped_column(Integer, nullable=True)
     url: Mapped[str] = mapped_column(String(512), nullable=False)
-    x1: Mapped[int] = mapped_column(Integer, nullable=False)
-    y1: Mapped[int] = mapped_column(Integer, nullable=False)
-    x2: Mapped[int] = mapped_column(Integer, nullable=False)
-    y2: Mapped[int] = mapped_column(Integer, nullable=False)
+    x1: Mapped[float] = mapped_column(Float, nullable=False)
+    y1: Mapped[float] = mapped_column(Float, nullable=False)
+    x2: Mapped[float] = mapped_column(Float, nullable=False)
+    y2: Mapped[float] = mapped_column(Float, nullable=False)
     create_time: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     update_time: Mapped[datetime] = mapped_column(DateTime, nullable=True)
